Add a simple script to run Phan checks
Script can be run as:
PHAN=/path/to/phan tests/phan/bin/phan
or if Phan is in $PATH simply with:
tests/phan/bin/phan

Note: Phan cannot be installed as a development dependency,
as long this library is compatible with PHP 5. Therefore,
Phan checks cannot also be run as a composer script
(a.k.a. "composer phan").